+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     -  Added Urban Bohman Method functionality to the SC Storm Runoff model
     -  South Carolina application selects nearest NHD stream and HUC 8 watershed and adds to report
     -  South Carolina application allows user to select a different NHD stream and HUC 8 watershed
+    -  Query NHD streams in South Carolina
+
   ### Changed
 
   ### Fixed